Corrects comparison type warnings in tests

+++ Test/BundleAgent/Bundle/CanonicalBlockTest.cpp
@@ -101,10 +101,10 @@ TEST(CanonicalBlockTest, GetMultipleLenght) {
   ASSERT_EQ(static_cast<uint64_t>(size1), calSize);
   rawBlock = rawBlock.substr(calSize);
   calSize = CanonicalBlock(rawBlock).getLength();
-  ASSERT_EQ(size2, calSize);
+  ASSERT_EQ(static_cast<uint64_t>(size2), calSize);
   rawBlock = rawBlock.substr(calSize);
   calSize = CanonicalBlock(rawBlock).getLength();
-  ASSERT_EQ(size3, calSize);
+  ASSERT_EQ(static_cast<uint64_t>(size3), calSize);
 }
 
 /**
