+++ Test/BundleAgent/Node/Neighbour/NeighbourDiscoveryTest.cpp
@@ -61,25 +61,25 @@ TEST(NeighbourDiscoveryTest, NeighbourCleanerTest) {
   NeighbourDiscovery nd(cf, nt);
   nt->update(std::make_shared<Neighbour>("node100", "192.168.1.1", 4000));
   auto neighbours = nt->getValues();
-  ASSERT_EQ(1, neighbours.size());
+  ASSERT_EQ(static_cast<size_t>(1), neighbours.size());
   nt->update(std::make_shared<Neighbour>("node101", "192.168.1.1", 4000));
   neighbours.clear();
   neighbours = nt->getValues();
-  ASSERT_EQ(2, neighbours.size());
+  ASSERT_EQ(static_cast<size_t>(2), neighbours.size());
   sleep(3);
   neighbours.clear();
   neighbours = nt->getValues();
-  ASSERT_EQ(2, neighbours.size());
+  ASSERT_EQ(static_cast<size_t>(2), neighbours.size());
   nt->update(std::make_shared<Neighbour>("node101", "192.168.1.1", 4000));
   sleep(2);
   neighbours.clear();
   neighbours = nt->getValues();
-  ASSERT_EQ(1, neighbours.size());
+  ASSERT_EQ(static_cast<size_t>(1), neighbours.size());
   ASSERT_EQ("node101", nt->getValue("node101")->getId());
   sleep(5);
   neighbours.clear();
   neighbours = nt->getValues();
-  ASSERT_EQ(0, neighbours.size());
+  ASSERT_EQ(static_cast<size_t>(0), neighbours.size());
   g_stop = true;
   // The neighbour cleaner thread has been stopped, so the new neighbours
   // must not be cleaned.
@@ -87,7 +87,7 @@ TEST(NeighbourDiscoveryTest, NeighbourCleanerTest) {
   sleep(5);
   neighbours.clear();
   neighbours = nt->getValues();
-  ASSERT_EQ(1, neighbours.size());
+  ASSERT_EQ(static_cast<size_t>(1), neighbours.size());
 }
 
 /**
@@ -120,10 +120,11 @@ TEST(NeighbourDiscoveryTest, NeighbourSendAndReceiveTest) {
   NeighbourDiscovery nd(cf, nt);
   sleep(3);
   auto neighbours = nt->getValues();
-  ASSERT_EQ(1, neighbours.size());
+  ASSERT_EQ(static_cast<size_t>(1), neighbours.size());
   ASSERT_EQ("node1", nt->getValue(*neighbours.begin())->getId());
   ASSERT_EQ("127.0.0.1", nt->getValue(*neighbours.begin())->getNodeAddress());
-  ASSERT_EQ(40000, nt->getValue(*neighbours.begin())->getNodePort());
+  ASSERT_EQ(static_cast<uint16_t>(40000),
+            nt->getValue(*neighbours.begin())->getNodePort());
   g_stop = true;
   sleep(5);
 }
